cleaned up unneeded conversions

+++ src/ft_write_x.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_write_x.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: juramos <juramos@student.42madrid.com>     +#+  +:+       +#+        */
+/*   By: juramos <juramos@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/28 11:52:28 by juramos           #+#    #+#             */
-/*   Updated: 2023/12/07 10:01:16 by juramos          ###   ########.fr       */
+/*   Updated: 2023/12/12 11:05:44 by juramos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ size_t	ft_write_x(va_list args)
 	size_t			count;
 
 	count = 0;
-	value = (unsigned int)va_arg(args, unsigned int);
+	value = va_arg(args, unsigned int);
 	count += putnbr_base_with_count_fd(value, "0123456789abcdef", 1);
 	return (count);
 }
@@ -29,7 +29,7 @@ size_t	ft_write_x_upper(va_list args)
 	size_t			count;
 
 	count = 0;
-	value = (unsigned int)va_arg(args, unsigned int);
+	value = va_arg(args, unsigned int);
 	count += putnbr_base_with_count_fd(value, "0123456789ABCDEF", 1);
 	return (count);
 }
